tests: Split out ZulipTestCase and WebhookTestCase to separate file
Fixes #1671
