+++ zerver/tests/test_events.py
@@ -52,7 +52,10 @@ from zerver.lib.actions import (
 
 from zerver.lib.event_queue import allocate_client_descriptor
 from zerver.lib.message import render_markdown
-from zerver.lib.test_helpers import ZulipTestCase, POSTRequestMock
+from zerver.lib.test_helpers import POSTRequestMock
+from zerver.lib.test_classes import (
+    ZulipTestCase,
+)
 from zerver.lib.validator import (
     check_bool, check_dict, check_int, check_list, check_string,
     equals, check_none_or, Validator
