+++ zerver/tests/test_messages.py
@@ -16,13 +16,16 @@ from zerver.lib.message import (
 )
 
 from zerver.lib.test_helpers import (
-    ZulipTestCase,
     get_user_messages,
     make_client,
     message_ids, message_stream_count,
     most_recent_message,
     most_recent_usermessage,
     queries_captured,
+)
+
+from zerver.lib.test_classes import (
+    ZulipTestCase,
 )
 
 from zerver.models import (
