+++ zerver/tests/test_realm_emoji.py
@@ -2,7 +2,7 @@
 from __future__ import absolute_import
 
 from zerver.lib.actions import get_realm, check_add_realm_emoji
-from zerver.lib.test_helpers import ZulipTestCase
+from zerver.lib.test_classes import ZulipTestCase
 from zerver.models import RealmEmoji
 import ujson
