+++ zerver/tests/webhooks/test_beanstalk.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from six import text_type
 from zerver.lib.webhooks.git import COMMITS_LIMIT
-from zerver.lib.test_helpers import WebhookTestCase
+from zerver.lib.test_classes import WebhookTestCase
 
 class BeanstalkHookTests(WebhookTestCase):
     STREAM_NAME = 'commits'
