+++ zerver/tests/webhooks/test_crashlytics.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from zerver.lib.test_helpers import WebhookTestCase
+from zerver.lib.test_classes import WebhookTestCase
 
 class CrashlyticsHookTests(WebhookTestCase):
     STREAM_NAME = 'crashlytics'
