+++ zerver/tests/webhooks/test_gitlab.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from zerver.lib.webhooks.git import COMMITS_LIMIT
-from zerver.lib.test_helpers import WebhookTestCase
+from zerver.lib.test_classes import WebhookTestCase
 
 class GitlabHookTests(WebhookTestCase):
     STREAM_NAME = 'gitlab'
