+++ zerver/tests/webhooks/test_transifex.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from six import text_type
 from typing import Any, Dict
-from zerver.lib.test_helpers import WebhookTestCase
+from zerver.lib.test_classes import WebhookTestCase
 
 class TransifexHookTests(WebhookTestCase):
     STREAM_NAME = 'transifex'
